[fix] mod BaseTechObject and move AttachedObjects to own class
